Final refactoring and getting quiz mode displaying
